requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-deepseek>=0.0.1
streamlit>=1.30.0
python-dotenv>=1.0.0
reportlab>=4.0.0
